+++ src/BreakoutNet/API/IBreakoutEvent.cs
@@ -1,0 +1,6 @@
+namespace BreakoutMods.BreakoutNet
+{
+    public interface IBreakoutEvent
+    {
+    }
+}