+++ src/BreakoutNet/API/IBreakoutModule.cs
@@ -1,0 +1,7 @@
+namespace BreakoutMods.BreakoutNet
+{
+    public interface IBreakoutModule
+    {
+        void Initialize(BreakoutModuleContext context);
+    }
+}